Restore the state of presenters in the backstack
Presenters in the backstack get paused when a new presenter gets pushed. This is by design, because they're in the background and should not consume resources. However, they also lose their state, because their composable function is no longer part of the composition.

`rememberSaveable()` provided by the Compose runtime allows you to save and restore state of composable functions. This PR adds a custom implementation for this mechanism, which allows us to return a `Model`, but is otherwise identical.