Support Back gestures for presenters.
This change introduces `BackGestureDispatcherPresenter`, which functions as a bridge between presenters and the UI layer to forward back gestures to presenters.

Two new functions `BackHandlerPresenter {}` and `PredictiveBackHandlerPresenter {}` were introduced that mimic the APIs from Compose Multiplatform `BackHandler {}` and `PredictiveBackHandler` to consume back gestures. The "Presenter" suffix was chosen to avoid conflicts and confusions with the APIs from Compose Multiplatform or Android Compose UI.

The UI layer integration will be implemented in subsequent changes.

See #57