Provide an easy to use back press testing API
Presenters that use `BackHandler {}` crash when calling `test()` on them in unit tests, because no `BackGestureDispatcherPresenter` is provided as composition local. A new test API `withBackGestureDispatcher()` wraps the presenter and provides a composition local. Back press events can be provided for tests using an optional `SharedFlow`.

This PR moves the `DefaultBackGestureDispatcherPresenter` to the `:public` module to make it reusable between the `:impl` and `:testing` module. There is no need to provide a different fake implementation for `BackGestureDispatcherPresenter`, because the real implementation is small, very efficient and no dependencies need to be replaced.