`@ContributesRenderer` code generator
Add a new KSP code generator that generates the code for `@ContributesRenderer` using Metro. The implementation is almost equivalent to the kotlin-inject implementation, except for that it's using Metro APIs. The big difference is that kotlin-inject implements map-multibindings by returning a `Pair<Abc, Def>` whereas Metro uses a `@ClassKey` annotation similar to Dagger 2. Therefore, we had to introduce the `@RendererKey` annotation specifically for Metro.

See #115