Fix and stop suppressing NPE when removing Android Views
In the past we were seeing this crash without being able to reproduce it:
```
java.lang.NullPointerException: Attempt to write to field 'android.view.ViewParent
    android.view.View.mParent' on a null object reference
      at android.view.ViewGroup.removeFromArray(ViewGroup.java:5384)
      at android.view.ViewGroup.removeViewInternal(ViewGroup.java:5581)
      at android.view.ViewGroup.removeViewInternal(ViewGroup.java:5543)
      at android.view.ViewGroup.removeView(ViewGroup.java:5474)
```
Because of that we simply caught and ignored the exception, but this caused more issues later at runtime.

We managed to reproduce this particular bug now and addressed the NPE properly. It no longer needs to be suppressed.

Fixes #136