Disallow changing the parent View for renderers
Until now it was allowed to change the parent view for a `Renderer` and this would remove and re-add the view at the new position. However, this scenario could lead to confusing results and race conditions. The new behavior is to create a unique `Renderer` for each parent and create distinct view hierarchies.

A side-effect of this change can be that parent `Renderers` embedding child `Renderers` now must actively remove views from the child `Renderers`, if the child `Renderer` position changed. If not, stale UI might be visible.

Helps with #136